Correct post published and modified datetimes
I got the original datetime values from the HTML. However it seems
that the offsets were not correct. This patch updates those values
with the "date_gmt" and "modified_gmt" values for each post from
the WordPress REST API.

It also seems that I had made an input error on the modDate for the
addthis-friendly-widget post. I have double-checked the value from
the API to confirm that it is now correct.