Create post excerpt system
This system has limitations imposed because it doesn't use Astro's
markdown system for rendering the post excerpt. The issue that I
ran into was that it doesn't support the `Image()` assets images.

This was a problem for the NAS part 3 post, which originally showed
a screenshot in the excerpt section. To work around this, I simply
moved the image below the excerpt section.