Explicitly `.sort()` articles list instead of using `.reverse()`
Astro has started showing the built (`npm run build`) version of
the articles list in the opposite order. When using the dev server
(`npm run dev`) the order is still correct. So it seems that I
cannot count on the order of Astro content collections.

To remedy this I have changed the ordering to use an explicit
`.sort()` by publish date instead of simply using `.reverse()`.
This will force the resulting list order to be consistent
regardless of the original content collection order.