Explicitly `.sort()` articles in RSS feed
This patch is to make the articles list ordering code in the RSS
feed consistent with the change to the main site.

This may have even been the root cause of the problem, since I was
not using `.toReversed()` to prevent modifying the list. However I
am unsure about that since the built version had the correct order
when this change was originally deployed.

Ref: Commit 5d716aed80383ed4eb95d654c87accc07a5dc24e