Remove `modDate` property from articles
This data is currently unused. If I do decide to surface a last
modified date somewhere then a manually updated piece of metadata
is probably not the best solution. I should be able to get that
information from git. Because everything is now managed in git and
all of the posts have been modified since the initial import.